Fix issue when initializing project from folder background
When clicking from a context menu on a folder background the exact directory
path was not forwarded to the initialize action. Therefore no content was
copied as a result. This is fixed now by including folder path from which user
had clicked on folder background.
Additionally, few logging points have been refactored to include more details.